LWDEV-6992 matching engine use big decimal

+++ src/main/kotlin/com/lykke/matching/engine/daos/LkkTrade.kt
@@ -1,11 +1,12 @@
 package com.lykke.matching.engine.daos
 
+import java.math.BigDecimal
 import java.util.Date
 
 data class LkkTrade(
         val assetPair: String,
         val clientId: String,
-        val price: Double,
-        val volume: Double,
+        val price: BigDecimal,
+        val volume: BigDecimal,
         val date: Date
 )
