+++ src/main/kotlin/com/lykke/matching/engine/notification/QuotesUpdate.kt
@@ -1,3 +1,5 @@
 package com.lykke.matching.engine.notification
 
-data class QuotesUpdate(val asset: String, val price: Double, val volume: Double)
+import java.math.BigDecimal
+
+data class QuotesUpdate(val asset: String, val price: BigDecimal, val volume: BigDecimal)
